Test coverage for if field is left blank
