+++ src/IntegrationTests/IntegrationTests.csproj
@@ -43,11 +43,17 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Services\MetricsProviderMockTests\FakeMetricsFileGetter.cs" />
+    <Compile Include="Services\MetricsProviderMockTests\when_creating_company_metrics_from_blank_value_in_BlankValue_File.cs" />
+    <Compile Include="Services\MetricsProviderMockTests\when_creating_company_metrics_from_invalid_weight_in_InvalidWeight_File.cs" />
     <Compile Include="Services\MetricsProviderMockTests\when_creating_company_metric_from_MetricTestFile.cs" />
     <Compile Include="SpecificationBase.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+    <None Include="TestFiles\BlankValue.csv">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Include="TestFiles\InvalidMetrics.csv">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
