+++ src/IntegrationTests/TestFiles/BlankValue.csv
@@ -1,0 +1,4 @@
+Name,IsDefault,Weight,AllowedAge,RequiresAlert
+,TRUE,0,,FALSE
+Metric 2,TRUE,0,1,FALSE
+Metric 3,TRUE,-50,1,FALSE
